Improve code quality checks and typing hygiene

+++ python/pdstools/app/decision_analyzer/_home_page.py
@@ -252,7 +252,7 @@ zoom, and hover for details.
                 )
 
                 row_count = raw_data.select(pl.len()).collect().item()
-                target_n = sample_kwargs.get("n", int(row_count * sample_kwargs.get("fraction", 1.0)))
+                target_n = int(sample_kwargs.get("n", int(row_count * sample_kwargs.get("fraction", 1.0))))
                 min_time, max_time = estimate_sampling_time(row_count, target_n)
                 time_msg = format_time_estimate(min_time, max_time)
                 sampling_msg += f" (estimated: {time_msg})"
